EICNET-750: Refactor EntityOperations to MessageCreator class per entity type.
