+++ lib/modules/eic_messages/src/Hooks/GroupContentMessageCreator.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace Drupal\eic_messages\Hooks;
+
+use Drupal\Core\Entity\EntityInterface;
+
+/**
+ * Class GroupContentMessageCreator.
+ */
+class GroupContentMessageCreator extends MessageCreatorBase {
+
+  /**
+   * Implements hook_group_content_insert().
+   *
+   * Sends out message notifications upon group content creation.
+   */
+  public function groupContentInsert(EntityInterface $entity) {
+    $messages = [];
+
+    /** @var \Drupal\group\Entity\GroupContent $entity */
+    $group_content_type = $entity->getGroupContentType();
+    if ($group_content_type->get('content_plugin') === 'group_membership') {
+      $relatedUser = $entity->getEntity();
+      $relatedGroup = $entity->getGroup();
+
+      // Prepare the message to the group owner.
+      $message = $this->entityTypeManager->getStorage('message')->create([
+        'template' => 'notify_new_member_joined',
+        'uid' => $relatedGroup->getOwnerId(),
+        'field_group_ref' => ['target_id' => $relatedGroup->id()],
+        'field_group_membership' => ['target_id' => $entity->id()],
+        'field_related_user' => ['target_id' => $relatedUser->id()],
+      ]);
+      $messages[] = $message;
+    }
+    $this->processMessages($messages);
+  }
+
+}
